import statements refactor: update package paths for previewLabApplication and LabPreview
