+++ core/src/jvmDev/kotlin/Main.kt
@@ -1,5 +1,5 @@
-import me.tbsten.compose.preview.lab.component.previewLabApplication
 import me.tbsten.compose.preview.lab.openfilehandler.OpenFileHandler
+import me.tbsten.compose.preview.lab.previewLabApplication
 import me.tbsten.compose.preview.lab.previewsForUiDebug
 
 fun main(): Unit = previewLabApplication(
