+++ integrationTest/app/src/androidMain/kotlin/me/tbsten/compose/preview/lab/sample/LabPreviewsSample.kt
@@ -5,8 +5,8 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
+import me.tbsten.compose.preview.lab.LabPreview
 import me.tbsten.compose.preview.lab.PreviewLab
-import me.tbsten.compose.preview.lab.component.LabPreview
 import me.tbsten.compose.preview.lab.field.StringField
 
 @LabPreview
